Moved LowPassFilter to shared, removed unnecessary usings.
