+++ Xamarin.Essentials/Compass/Compass.shared.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Threading;
 
 namespace Xamarin.Essentials
 {
